+++ entry/models.py
@@ -1452,6 +1452,20 @@ class Entry(ACLBase):
         super(Entry, self).__init__(*args, **kwargs)
         self.objtype = ACLObjType.Entry
 
+    def add_alias(self, name):
+        # validate name that is not duplicated with other Item names and Aliases in this model
+        if not self.schema.is_available(name):
+            raise ValueError("Specified name has already been used by other Item or Alias")
+
+        return AliasEntry.objects.create(name=name, entry=self)
+
+    def delete_alias(self, name):
+        alias = AliasEntry.objects.filter(
+            name=name, entry__schema=self.schema, entry__is_active=True
+        ).first()
+        if alias:
+            alias.delete()
+
     def add_attribute_from_base(self, base: EntityAttr, request_user: User):
         if not isinstance(base, EntityAttr):
             raise TypeError('Variable "base" is incorrect type')
@@ -2421,3 +2435,14 @@ class AdvancedSearchAttributeIndex(models.Model):
                 return self.raw_value
             case _:
                 print("TODO implement it")
+
+
+class AliasEntry(models.Model):
+    name = models.CharField(max_length=200)
+
+    # This indicates alias of this Entry
+    entry = models.ForeignKey(
+        Entry,
+        related_name="aliases",
+        on_delete=models.CASCADE,
+    )