+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## In development
 
 ### Added
+* Introduced a new feature to identify Item using another names that are related with
+  specific Item (it is named as Alias).
+  Contributed by @hinashi, @userlocalhost
 
 ### Changed